stat: move Boot func to a separate file
